refactor: limit select 분리
